feature: handle bundle response
Introduce a promise in case of bundle queries which gets resolved and invokes the callback  once all the enteries are processed rather than per entry.
In case of successful bundle request the callback would be invoked with array of successful enteries.
In case of failure the callback would be invoked with a Message containing the details of the failure