fix: single filter or sorter object
According to ui5 ListBinding the intial application filters can be a single filter or an array of filters https://openui5.hana.ondemand.com/api/sap.ui.model.ListBinding
aFilters?
sap.ui.model.Filter
sap.ui.model.Filter[]
Predefined filter/s (can be either a filter or an array of filters)
So if a single filter object is given instead of array it doesnt get applied since the model expects an array as default. This fix ensures to convert single instance of filter to an array of filters before applying it to the corresponding binding. oData ListBinding also handles it in the same way