fix: complex filter query (#267)
* fix: for when value type of string is not present

In case of mdc controls the filter object is generated by ui5 and hence the valuetype for string is not available to determine if it needs to be enclosed in quotes. So this pr aims to better identify string type based on js types and enclose it in quotes. if a valid date is given in string it wont be enclosed. Also removing the conversion to iso string since if the type of the resource and not date time this would lead to incorrect filter results. For.eg. birthDate is of type date. This is a short term solution to handle dates. The long term solution of having Fhir date types and handling it accordingly will be done in a later stage

* test: add units